refactor: 헤더 상태 관련된 코드 WikiHeader 내부로 이동
